:boom: change column name of mine-sweeper-records table
